--- src/Components/index.ts
@@ -1,4 +1,0 @@
-export { default as EditableTable } from './EditableTable';
-export type { ExportColumn } from './ExportButton';
-export { default as UploadFile } from './UploadFile';
-export { default as Header } from './Header';
