+++ src/components/export/ExportButton.tsx
@@ -1,6 +1,6 @@
 import { DocumentArrowDownIcon, DocumentMagnifyingGlassIcon } from '@heroicons/react/24/outline';
-import { getPdfFilename, generatePDF } from '@exportbutton';
-import type { ExportColumn } from '@exportbutton';
+import { getPdfFilename, generatePDF } from '@components/export';
+import type { ExportColumn } from '@components/export';
 
 interface ExportButtonProps<T extends Record<string, unknown> & { SheetName?: string }> {
   data: T[];
