+++ src/components/table/index.ts
@@ -1,0 +1,3 @@
+export { default as EditableTable } from './EditableTable';
+export * from './tableTypes';
+export * from './tableUtils';
