+++ src/components/table/tableTypes.ts
@@ -1,0 +1,4 @@
+export interface EditableTableProps {
+  data: Record<string, unknown>[];
+  dataOnChange?: (newData: Record<string, unknown>[]) => void;
+}
