+++ src/components/upload/index.ts
@@ -1,0 +1,3 @@
+export { default as UploadFile } from './UploadFile';
+export * from './uploadTypes';
+export * from './uploadUtils';
