+++ src/components/upload/uploadTypes.ts
@@ -1,0 +1,7 @@
+export type RowData = {
+  [key: string]: string | number | boolean | null;
+};
+
+export interface UploadFileProps {
+  onDataParsed: (data: { [sheetName: string]: RowData[] }, fileName: string) => void;
+}
