+++ src/hooks/index.ts
@@ -1,1 +1,2 @@
 export { default as useLocalSheet } from './useLocalSheet';
+export * from './useSheetColumns';
