+++ src/types/excel.d.ts
@@ -1,0 +1,7 @@
+export type ExcelRow = {
+  [key: string]: string | number | boolean | null;
+};
+
+export type SheetData = {
+  [sheetName: string]: ExcelRow[];
+};
