+++ tsconfig.app.json
@@ -27,11 +27,15 @@
     /* Aliases */
     "baseUrl": "./src",
     "paths": {
-      "@styles": ["styles"],
-      "@components": ["Components"],
       "@src": ["."],
-      "@exportbutton": ["Components/ExportButton"],
-      "@hooks": ["hooks"]
+      "@styles": ["styles"],
+      "@components": ["components"],
+      "@components/export": ["components/export"],
+      "@components/layout": ["components/layout"],
+      "@components/table": ["components/table"],
+      "@components/upload": ["components/upload"],
+      "@hooks": ["hooks"],
+      "@types": ["types"]
     }
   },
   "include": ["src"]
